+++ girepository/gifunctioninfo.h
@@ -87,9 +87,9 @@ typedef enum
 GI_AVAILABLE_IN_ALL
 gboolean              gi_function_info_invoke         (GIFunctionInfo    *info,
                                                        const GIArgument  *in_args,
-                                                       int                n_in_args,
+                                                       gsize              n_in_args,
                                                        const GIArgument  *out_args,
-                                                       int                n_out_args,
+                                                       gsize              n_out_args,
                                                        GIArgument        *return_value,
                                                        GError           **error);
